提取controller逻辑到helper
